+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  * add command export-packages
  * add polling state of the sent command
  * add ansible module to output the log of the sent command to stdout
+ * add ansible playbook for offline-snapshot and offline-compaction-snapshot
 
 
 ### 1.1.0
@@ -12,4 +13,3 @@
 
 ### 1.0.0
 * Initial version
-